Update forgot password form for layered architecture
