+++ src/app/(auth-pages)/forgot-password/page.tsx
@@ -1,14 +1,11 @@
-import { headers } from 'next/headers'
 import Link from 'next/link'
-import { redirect } from 'next/navigation'
 
-import { authActions } from '@/app/_actions/auth-actions'
 import Field from '@/components/form/field'
 import { FormMessage, Message } from '@/components/form/form-message'
 import { SubmitButton } from '@/components/submit-button'
 import { Input } from '@/components/ui/input'
-import { encodedRedirect } from '@/utils/encoded-redirect'
-import { tryCatch } from '@/utils/try-catch'
+
+import { forgotPasswordFormAction } from './forgot-password-form-action'
 
 export default async function ForgotPassword(props: { searchParams: Promise<Message> }) {
   const searchParams = await props.searchParams
@@ -32,27 +29,4 @@ export default async function ForgotPassword(props: { searchParams: Promise<Mess
       </form>
     </>
   )
-}
-
-async function forgotPasswordFormAction(formData: FormData) {
-  'use server'
-  const email = formData.get('email')?.toString()
-  const origin = (await headers()).get('origin')
-  const callbackUrl = formData.get('callbackUrl')?.toString()
-
-  if (!email) {
-    return encodedRedirect('error', '/forgot-password', 'Email is required')
-  }
-
-  const { error } = await tryCatch(authActions.forgotPassword({ email, origin }))
-
-  if (error) {
-    return encodedRedirect('error', '/forgot-password', 'Could not reset password')
-  }
-
-  if (callbackUrl) {
-    return redirect(callbackUrl)
-  }
-
-  return encodedRedirect('success', '/forgot-password', 'Check your email for a link to reset your password.')
 }
